index.html and bare.html are now separate pages that embed friendlycode.
Specifically, index.html assumes embedding within the Thimble website
and features Thimble-specific branding, while bare.html is a trivial
embedding of friendlycode.

The many stylesheets used by both pages are now consolidated into
friendlycode.css, which (for now) just @imports all the stylesheets.
A production build system can compile these into a single file for
better performance.
