+++ bare.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html class="deployment-type-development">
+  <head>
+    <meta charset="utf-8">
+    <base target="_blank">
+
+    <title>Barebones Friendlycode Editor</title>
+    <link rel="stylesheet" href="css/friendlycode.css">
+  </head>
+  <body>
+    <div id="bare-fc-holder" class="friendlycode-loading"></div>
+
+    <script src="js/require-config.js"></script>
+    <script src="js/require.min.js"></script>
+    <script>
+    require(["jquery", "friendlycode"], function($, FriendlycodeEditor) {
+      return FriendlycodeEditor({
+        publishURL: "https://webpagemaker-dev.allizom.org",
+        container: $("#bare-fc-holder")
+      });
+    });
+    </script>
+  </body>
+</html>
