+++ css/editor.css
@@ -30,14 +30,6 @@ img {
     border: 0 none;
 }
 
-#friendlycode-holder {
-  position: absolute;
-  top: 79px;
-  bottom: 0;
-  left: 0;
-  right: 0;
-}
-
 .friendlycode-loading > * {
   visibility: hidden;
 }
@@ -110,36 +102,6 @@ img {
   border-bottom: 1px solid #D6D6D6;
   
   background: white;
-}
-
-#header {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  height: 79px;
-
-  background: #f5f5f5;
-  background-image: url("../img/header-gradient.png");
-  background-position: 0 0;
-  -webkit-box-shadow: 0 5px 5px 0 rgba(233,233,233,0.5);
-  box-shadow: 0 5px 5px 0 rgba(233,233,233,0.5);
-  border-top: 2px solid white;
-  
-  z-index: 2;
-}
-
-#header .logo {
-  display: inline-block;
-  font-family: OpenSansLight, sans-serif;
-  margin-left: 15px;
-  width: 200px;
-  margin-top: 10px;
-}
-
-#header .logo a {
-  color: #27AAE1;
-  text-decoration: none;
 }
 
 .nav-options {
@@ -258,9 +220,6 @@ img {
   }
   .hints-nav-item {
     left: 180px !important;
-  }
-  #tabzilla {
-    right: 16% !important;
   }
 }
