+++ css/friendlycode.css
@@ -1,0 +1,19 @@
+/* 
+ * This stylesheet is only intended for development purposes,
+ * due to the inefficiency of @import [1]. For production use, this
+ * file should be a raw concatenation of all the @import rules below.
+ *
+ * [1] http://www.stevesouders.com/blog/2009/04/09/dont-use-import/
+ */
+
+@import url(../codemirror2/lib/codemirror.css);
+
+@import url(jsbin-codemirror-theme.css);
+@import url(tipsy.css);
+@import url(ubuntumono/stylesheet.css);
+@import url(opensans/stylesheet.css);
+@import url(opensymbolcropped/stylesheet.css);
+@import url(editor.css);
+@import url(errorhelp.css);
+@import url(modals.css);
+
